Add extra tests around CLI options
